+++ index.js
@@ -3,5 +3,5 @@ function debugEnabled() {
   return window.localStorage.embettyDebug === 'true'
 }
 if (debugEnabled()) {
-  console.log('embetty-version: ' + require('./package.json').version)
+  console.log(`embetty-version: ${require('./package.json').version}`)
 }