+++ polyfills.js
@@ -5,7 +5,9 @@ require('whatwg-fetch')
 try {
   /* eslint no-eval: "warn" */
   eval(require('@webcomponents/webcomponentsjs/custom-elements-es5-adapter.js'))
-} catch (error) {}
+} catch (error) {
+  console.log(error)
+}
 
 require('@webcomponents/webcomponentsjs')
 require('intersection-observer')