+++ webpack.config.js
@@ -8,18 +8,16 @@ const {
   version,
 } = require('./package.json')
 
-const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin
+const BundleAnalyzerPlugin = require('webpack-bundle-analyzer')
+  .BundleAnalyzerPlugin
 const path = require('path')
 const webpack = require('webpack')
 
-const prod = process.argv.indexOf('-p') !== -1
+const prod = process.argv.includes('-p')
 
 module.exports = {
   entry: {
-    embetty: [
-      './polyfills.js',
-      './index.js'
-    ]
+    embetty: ['./polyfills.js', './index.js'],
   },
   output: {
     path: path.resolve(__dirname, 'dist'),
@@ -32,36 +30,27 @@ module.exports = {
       {
         test: /\.js$/,
         exclude: /custom-elements-es5-adapter/,
-        use: [
-          'cache-loader',
-          'babel-loader'
-        ]
+        use: ['cache-loader', 'babel-loader'],
       },
       {
         test: /\.scss$/,
-        use: [
-          'css-loader',
-          'postcss-loader',
-          'sass-loader'
-        ]
+        use: ['css-loader', 'postcss-loader', 'sass-loader'],
       },
       {
         test: /^custom-elements-es5-adapter\.js$/,
-        use: [
-          'raw-loader'
-        ]
-      }
-    ]
+        use: ['raw-loader'],
+      },
+    ],
   },
   plugins: [
     new webpack.IgnorePlugin(/vertx/), // see https://github.com/parcel-bundler/parcel/issues/141
     new BundleAnalyzerPlugin({ analyzerMode: 'static', openAnalyzer: false }),
     new webpack.BannerPlugin({
-      banner: `${title || name} - v${version} - ${(new Date()).toGMTString()}
+      banner: `${title || name} - v${version} - ${new Date().toGMTString()}
 ${homepage}
-Copyright (c) ${(new Date()).getFullYear()} Heise Medien GmbH & Co. KG
+Copyright (c) ${new Date().getFullYear()} Heise Medien GmbH & Co. KG
 Contributors: ${author.name}, ${contributors.map(c => c.name).join(', ')}
-Licensed under the ${license} license`
-    })
+Licensed under the ${license} license`,
+    }),
   ],
 }